+++ src/engine/strat_engine/backstore/setup.rs
@@ -15,10 +15,8 @@ use devicemapper::{devnode_to_devno, Device};
 
 use stratis::{ErrorEnum, StratisError, StratisResult};
 
-use super::super::super::structures::Table;
-use super::super::super::types::{Name, PoolUuid};
+use super::super::super::types::PoolUuid;
 
-use super::super::pool::{check_metadata, StratPool};
 use super::super::serde_structs::{BackstoreSave, PoolSave};
 
 use super::blockdev::StratBlockDev;
@@ -26,63 +24,6 @@ use super::device::{blkdev_size, is_stratis_device};
 use super::metadata::BDA;
 use super::util::get_stratis_block_devices;
 
-/// Setup a pool from constituent devices in the context of some already
-/// setup pools. Return an error on anything that prevents the pool
-/// being set up.
-pub fn setup_pool(
-    pool_uuid: PoolUuid,
-    devices: &HashMap<Device, PathBuf>,
-    pools: &Table<StratPool>,
-) -> StratisResult<(Name, StratPool)> {
-    // FIXME: In this method, various errors are assembled from various
-    // sources and combined into strings, so that they
-    // can be printed as log messages if necessary. Instead, some kind of
-    // error-chaining should be used here and if it is necessary
-    // to log the error, the log code should be able to reduce the error
-    // chain to something that can be sensibly logged.
-    let info_string = || {
-        let dev_paths = devices
-            .values()
-            .map(|p| p.to_str().expect("Unix is utf-8"))
-            .collect::<Vec<&str>>()
-            .join(" ,");
-        format!("(pool UUID: {}, devnodes: {})", pool_uuid, dev_paths)
-    };
-
-    let metadata = get_metadata(pool_uuid, devices)?.ok_or_else(|| {
-        let err_msg = format!("no metadata found for {}", info_string());
-        StratisError::Engine(ErrorEnum::NotFound, err_msg)
-    })?;
-
-    if pools.contains_name(&metadata.name) {
-        let err_msg = format!(
-            "pool with name \"{}\" set up; metadata specifies same name for {}",
-            &metadata.name,
-            info_string()
-        );
-        return Err(StratisError::Engine(ErrorEnum::AlreadyExists, err_msg));
-    }
-
-    check_metadata(&metadata)
-        .or_else(|e| {
-            let err_msg = format!(
-                "inconsistent metadata for {}: reason: {:?}",
-                info_string(),
-                e
-            );
-            Err(StratisError::Engine(ErrorEnum::Error, err_msg))
-        })
-        .and_then(|_| {
-            StratPool::setup(pool_uuid, devices, &metadata).or_else(|e| {
-                let err_msg = format!(
-                    "failed to set up pool for {}: reason: {:?}",
-                    info_string(),
-                    e
-                );
-                Err(StratisError::Engine(ErrorEnum::Error, err_msg))
-            })
-        })
-}
 
 /// Find all Stratis devices.
 ///
